x11 improvements

+++ WDL/swell2/swell-misc.cpp
@@ -1450,8 +1450,8 @@ void GetCursorPos(POINT *pt)
 #ifdef SWELL_TARGET_SDL3
   float mx, my;
   SDL_GetGlobalMouseState(&mx, &my);
-  pt->x = swell_log_to_phys((int)mx);
-  pt->y = swell_log_to_phys((int)my);
+  pt->x = swell_log_to_phys(mx);
+  pt->y = swell_log_to_phys(my);
 #else
   pt->x = 0; pt->y = 0;
 #endif
@@ -1462,7 +1462,7 @@ DWORD GetMessagePos()
 #ifdef SWELL_TARGET_SDL3
   float mx, my;
   SDL_GetGlobalMouseState(&mx, &my);
-  return MAKELPARAM(swell_log_to_phys((int)mx), swell_log_to_phys((int)my));
+  return MAKELPARAM(swell_log_to_phys(mx), swell_log_to_phys(my));
 #else
   return 0;
 #endif
